web-animations/animation-group/animation-group-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup constructor test</title>
<meta name="assert" content="Creates a new AnimationGroup object">
<link rel="help" href="http://w3c.github.io/web-animations/#the-animationgroup-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var group = new AnimationGroup(null);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    assert_object_equals(group.timing, DEFAULT_TIMING, 'AnimationGroup timing members ' +
        'should be set to default values');
}, 'Test AnimationGroup(null) constructor creates AnimationGroup instance ' +
    'without children and with default timing');

test(function() {
    var group = new AnimationGroup([]);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    assert_object_equals(group.timing, DEFAULT_TIMING, 'AnimationGroup timing members ' +
        'should be set to default values');
}, 'Test AnimationGroup constructor with emplty child list and with no second argument ' +
    'creates AnimationGroup instance without children and with default timing');

test(function() {
    var group = new AnimationGroup([], 5);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    var timing = Object.create(DEFAULT_TIMING);
    timing.duration = 5;
    assert_object_equals(group.timing, timing, 'AnimationGroup timing members ' +
        'should be set to default values, except duration');
}, 'Test AnimationGroup constructor with emplty child list and with double second argument ' +
    'creates AnimationGroup instance without children and with default timing, except duration');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var node = new AnimationGroup([], timing);

    assert_array_equals(node.children, [], 'AnimationGroup should have no children');
    assert_object_equals(node.timing, timing, 'AnimationGroup timing should be equal ' +
        'to the value passed to the constructor');
}, 'Test AnimationGroup constructor with emplty child list and with AnimationTimingInput ' +
    'second argument creates AnimationGroup instance without children and with given timing');

test(function() {
    var node = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var children = [node];
    var group = new AnimationGroup(children);

    assert_array_equals(group.children, children, 'AnimationGroup should have a single child');
}, 'Test AnimationGroup constructor creates AnimationGroup instance with a single child');

test(function() {
    var node1 = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var node2 = new AnimationGroup([]);
    var node3 = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var children = [node1, node2, node3];
    var group = new AnimationGroup(children);

    assert_array_equals(group.children, children, 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
}, 'Test AnimationGroup constructor creates AnimationGroup instance with children');

test(function() {
    var node = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var group = new AnimationGroup([node, node]);

    assert_array_equals(group.children, [node], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test AnimationGroup constructor removes Animation node from its parent group ' +
    'before inserting to child list even this is the same group');

test(function() {
    var node = new AnimationGroup([]);
    var group = new AnimationGroup([node, node]);

    assert_array_equals(group.children, [node], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test AnimationGroup constructor removes AnimationGroup node from its parent ' +
    'group before inserting to child list even this is the same group');

test(function() {
    var node = new AnimationSequence([]);
    var group = new AnimationGroup([node, node]);

    assert_array_equals(group.children, [node], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test AnimationGroup constructor removes AnimationSequence node from its parent ' +
    'group before inserting to child list even this is the same group');

test(function() {
    var node = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var group1 = new AnimationGroup([node]);
    var group2 = new AnimationGroup([node]);

    assert_array_equals(group1.children, [], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(group2.children, [node], 'AnimationGroup constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationGroup constructor removes Animation node from its parent ' +
    'group before inserting to child list');

test(function() {
    var node = new AnimationGroup([]);
    var group1 = new AnimationGroup([node]);
    var group2 = new AnimationGroup([node]);

    assert_array_equals(group1.children, [], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(group2.children, [node], 'AnimationGroup constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationGroup constructor removes AnimationGroup node from its parent ' +
    'group before inserting to child list');

test(function() {
    var node = new AnimationSequence([]);
    var group1 = new AnimationGroup([node]);
    var group2 = new AnimationGroup([node]);

    assert_array_equals(group1.children, [], 'AnimationGroup constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(group2.children, [node], 'AnimationGroup constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationGroup constructor removes AnimationSequence node from its parent ' +
    'group before inserting to child list');
</script>
</body>
